+++ weblate/trans/models/unit.py
@@ -190,6 +190,17 @@ class UnitQuerySet(models.QuerySet):
     def order(self):
         return self.order_by("-priority", "position")
 
+    def filter_access(self, user):
+        if user.is_superuser:
+            return self
+        return self.filter(
+            Q(translation__component__project_id__in=user.allowed_project_ids)
+            & (
+                Q(translation__component__restricted=False)
+                | Q(translation__component_id__in=user.component_permissions)
+            )
+        )
+
 
 class Unit(models.Model, LoggerMixin):